fix(parser): Fix CI errors after byte conversion
- Remove unnecessary `as u8` cast in ECMAScript lexer (clippy error)
- Fix XML parser to use `cur_as_char()` and byte literals for char comparisons
- Fix HTML parser current_char usage (already in staged changes)

+++ crates/swc_ecma_parser/src/lexer/mod.rs
@@ -2198,7 +2198,7 @@ impl<'a> Lexer<'a> {
     fn read_str_lit(&mut self) -> LexResult<Token> {
         debug_assert!(self.cur() == Some(b'\'') || self.cur() == Some(b'"'));
         let start = self.cur_pos();
-        let quote = self.cur().unwrap() as u8;
+        let quote = self.cur().unwrap();
 
         self.bump(); // '"' or '\''
